Move logic from app to workouts page
